// apps/mobile/src/components/ui/toast/types.ts
type PromiseOptions = {
	promise: Promise<unknown>;
	success: (result: any) => string; // TODO: type this with generics
	error: ((error: unknown) => string) | string;
	loading: string;
};

export type ToastPosition = "top-center" | "bottom-center" | "center";

export type ToastTheme = "light" | "dark" | "system";

export type ToastSwipeDirection = "left" | "up";

export type ToastVariant = "success" | "error" | "warning" | "info" | "loading";

export type AutoWiggle = "never" | "toast-change" | "always";

export type ToastAction = {
	label: string;
	onClick: () => void;
};

export type ToastProps = StyleProps & {
	id: string | number;
	title: string;
	variant: ToastVariant;
	jsx?: React.ReactNode;
	description?: string;
	invert?: boolean;
	important?: boolean;
	duration?: number;
	position?: ToastPosition;
	dismissible?: boolean;
	icon?: React.ReactNode;
	action?: ToastAction | React.ReactNode;
	cancel?: ToastAction | React.ReactNode;
	close?: React.ReactNode;
	closeButton?: boolean;
	richColors?: boolean;
	onDismiss?: (id: string | number) => void;
	onAutoClose?: (id: string | number) => void;
	promiseOptions?: PromiseOptions;
	actionButtonStyle?: ViewStyle;
	actionButtonTextStyle?: TextStyle;
	cancelButtonStyle?: ViewStyle;
	cancelButtonTextStyle?: TextStyle;
	onPress?: () => void;
};

export type ToastRef = {
	wiggle: () => void;
};

export function isToastAction(
	action: ToastAction | React.ReactNode,
): action is ToastAction {
	return (action as ToastAction)?.onClick !== undefined;
}

type ExternalToast = Omit<
	ToastProps,
	"id" | "type" | "title" | "jsx" | "promise" | "variant"
> & {
	id?: string | number;
};
// ...
export declare const toast: ((
	message: string,
	data?: ExternalToast,
) => string | number) & {
	success: (message: string, data?: ExternalToast) => string | number;
	info: (message: string, data?: ExternalToast) => string | number;
	error: (message: string, data?: ExternalToast) => string | number;
	warning: (message: string, data?: ExternalToast) => string | number;
	custom: (jsx: React.ReactElement, data?: ExternalToast) => string | number;
	promise: <T>(
		promise: Promise<T>,
		options: Omit<PromiseOptions, "promise">,
	) => string | number;
	loading: (message: string, data?: ExternalToast) => string | number;
	dismiss: (id?: string | number) => string | number | undefined;
	wiggle: (id: string | number) => void;
};